feat: #44 위시리스트 상품 조회, 삭제 구현

+++ Android/FundYou/data/src/main/java/com/ssafy/fundyou/data/remote/repository/WishListRepositoryImpl.kt
@@ -21,7 +21,7 @@ internal class WishListRepositoryImpl @Inject constructor(private val wishListRe
         return wishListRemoteDataSource.addWishListItem(request).statusCode
     }
 
-    override suspend fun deleteWishListItem(itemId: Int) : Int {
+    override suspend fun deleteWishListItem(itemId: Long) : Int {
         return wishListRemoteDataSource.deleteWishListItem(itemId).statusCode
     }
 }
